REFACTOR/EXTRACT/include to library
- 3 functions moved to library
