+++ pracs/01-a_tutorial_introduction/10-External_var_n_scope/srcs/next_column.c
@@ -1,0 +1,21 @@
+#include <unistd.h>
+
+ssize_t next_column(const char *src)
+{
+	ssize_t	i;
+	char ch;
+
+	i = 0;
+	
+	while ((ch = *(src + i)) && ch != '.' && ch != '\n')
+		i++;
+
+	if (ch == '.') {
+		while ((ch = *(src + i)) != '.')
+			i++;
+	} else
+		i = -1;
+
+	return (i);
+}
+
