+++ pracs/01-a_tutorial_introduction/10-External_var_n_scope/srcs/next_nonblank.c
@@ -1,0 +1,18 @@
+#include <unistd.h>
+
+ssize_t next_nonblank(const char *src)
+{
+	ssize_t	i;
+	char	ch;
+
+	i = 0;
+
+	while ((ch = *(src + i))) {
+		if (ch != ' ' || ch != '\t' || ch != '\n')
+			return (i);
+		i++;
+	}
+
+	return (-1);
+}
+
